Start intake roller earlier
