+++ src/main/java/frc/robot/subsystems/Intake.java
@@ -21,7 +21,7 @@ public class Intake extends SmartSubsystemBase {
     private static final double DEPLOY_EXTEND_SPEED = 0.3;
     private static final double DEPLOY_RETRACT_SPEED = -0.3;
 
-    private static final double ROLLER_THRESHOLD = 10.0;
+    private static final double ROLLER_THRESHOLD = 8.0;
 
     private final Modifier limit;
 
@@ -51,12 +51,12 @@ public class Intake extends SmartSubsystemBase {
             // Using validators in a modifier in combination with using it to stop the
             // command
             deploymentMotor.set(limit.applyAsDouble(DEPLOY_EXTEND_SPEED));
-            // if (deploymentMotor.getEncoder().getDistance() >= ROLLER_THRESHOLD) {
-            // rollerMotor.set(rollerDirection.apply(ROLLER_SPEED));
-            // }
+            if (deploymentMotor.getEncoder().getDistance() >= ROLLER_THRESHOLD) {
+                rollerMotor.set(rollerDirection.apply(ROLLER_SPEED));
+            }
         }).runsUntil(deploymentMotor::errored).onEnd((interrupted) -> {
             deploymentMotor.set(0.0);
-            rollerMotor.set(rollerDirection.apply(ROLLER_SPEED));
+            // rollerMotor.set(rollerDirection.apply(ROLLER_SPEED));
         });
     }
 
@@ -93,6 +93,7 @@ public class Intake extends SmartSubsystemBase {
     public void periodic() {
         SmartDashboard.putNumber("Intake Current Draw 1 (amps)", current.getAsDouble());
         SmartDashboard.putNumber("Intake Current Draw 2 (amps)", current2.getAsDouble());
+        SmartDashboard.putNumber("Intake Deploy Encoder", deploymentMotor.getEncoder().getDistance());
     }
 
     @Override
